.github/workflows/scripts/validation/lint-gamedb/requirements.txt
PyYAML==5.4.1